Expanded the funcionality of uuid.
The default behavior is the same, but now it takes a type parameter,
which currently only supports NIL and RANDOM, but will be expanded later
to support all 5 versions of UUIDs. The documentation is more detailed,
and there are examples now, and it is set up to easily return the other
types of UUIDs later.