Add a way for objects to provide a static constructor
Previously, in NativeTypeList, there was no way to customize how an invalid
instance was constructed. Now, if a subclass of Mixed provides a method named
ConstructInvalidInstance, this will be called instead. For now, this is not
necessary in any native classes, but since this picks up extension objects
as well, it might be useful already.