Add ValueType semantics
This allows objects to mark themselves as value types. Currently, all
primitives are marked as such. In the future, the compiler may be able
to further optimize these types, particularly if true compilation is
added. For now, this is unused, and the compiler treats everything as
pass by reference, but this is not ideal. At least going forward now,
the ValueType class is established, and other classes can go ahead and
start marking themselves as such. Eventually, this mechanism can
probably be exposed to user classes, but for now it is internal only.