Properly pass column information through the exception system
Previously, the causedBy section of the exceptions were not properly
printed in the default unhandled exception handler. Now, the column
information that is passed through is properly printed (though I think
the column information itself is still not accurate, but this is a
problem in the compiler, not the exception system). Also added "at" to
the beginning of each stack trace line, which brings the format more in
line with other languages stacktrace prints.