Fix site-deploy
Now, in addition to fixing the actual issue (in a way that that won't happen again),
we now also print out usable error messages. For whatever reason, Logger was not
working, so instead of fixing that, it has simply been removed, and made to work with
the existing paradigm anyways. This should make it easier to diagnose errors within
the documentation itself in the future.