Work on making clamp work as a composite function
This is a good example of a function for this, as it has no external
dependencies, is purely a math based function, has a simple
implementation, but takes full advantage of the optimization system.

There are also improvements to the ArgumentParser, along with a few bug
fixes in that area.