Add build-extension tool to cmdline tools
This allows for systems with git and maven installed to quickly and
easily pull down and build extensions. Assuming the extension is
properly set up with git and maven, the only need is to provide the
clone url to the tool. The tool will then git clone the repo, and use
maven to build it, and finally, copy the artifact into the chose
extension directory, defaulting to the current installation's directory.

This commit also fixes a couple of bugs. First, there was a potential
memory leak in CommandExecutor, as well as the fact that it closed
System.out/err. This would, I believe, have minor impact, as only a few
things would have used the buggy modes, but it did affect the
site-deploy tool, as well as shell_adv.

Secondly, when running in Java version 9 or above, every startup would
warn of reflective access. When launching MethodScript with the mscript
shortcut, it now provides the proper command line parameters to suppress
this warning. This would not help when using the jar directly, and
anyways is not a permanent fix, but in the meantime should be less
annoying until Java 9+ is officially supported.