Native clamp (#513)
Make clamp() work as a composite function

Moving forward, this is the most ideal way to code general purpose
functions that don't specifically require platform support. This should
make these functions portable if the backing engine is changed from
Java to something else, as well as once things like compilation to native
are supported, etc, these will be automatically picked up once the
underlying functions are all supported.

clamp() is a good example of a function for this, as it has no external
dependencies, is purely a math based function, has a simple
implementation, but takes full advantage of the optimization system.

There are also improvements to the ArgumentParser, along with a few bug
fixes in that area.