Begin work on creating the ObjectTable object chain.
This will eventually be the underlying mechanism which holds the
defintion of all objects. The first step is to create the structure, the
second step is to be able to fill it in with native classes. There are
several other steps after this to make it useable, but these are the
steps necessary to make it complete.

After it is complete, the next step will probably be to just allow
native types to provide properties, but this will require quite a bit of
work, especially when it comes to ensuring that all the necessary info
is available through annotations, and parameters are all compatible with
Mixed, etc. However, finishing the ObjectTable will be a great first
step.