Add ObjectHelpers. Support generic parameters in NonInheritsImplements.
Begin work on a POJOConversion framework, which will allow the exposed
methods in a native object to use POJOs, while transparently in the
background, the values are converted to and from Mixed values for use in
user code. This should make the code much easier to write, and much
faster as well. It may also assist in the future if there is a need to
map third party classes into the ecosystem.