Wrote more on topics, write some of the standard library.
Several more sections have been added to the objects design doc, as well
as a small section about manifests and annotations. In addition to this,
a new cmdline tool, new-type has been added. This allows quick and
correct creation of classes (and types) to be created. Additionally,
many classes in the standard library have been added, though they are
not yet compilable, since the compiler does not support this. However,
they serve as a great base to begin working against for the compiler
changes, and also to see larger segments of code using the designs
described in the design doc. There may be some (unintentional) compiler
errors in the code, but those will be identified later in the process.

Primarily the MethodScript code that is written focuses on supporting
the manifest file, but several pieces of supporting code have also been
written, all primarily annotations. The code is subject to be moved
around and refactored though.

Additionally, a workspace has been created for Visual Studio Code to
more easily edit the methodscript directory.