Object table (#517)
I am going ahead and merging this in to master, as it will be too difficult to manage two different branches for that long. So instead, I am going to do a slow rollout, while making some features available sooner than others, though until the features are all ready, they should not be used, and are subject to complete change. "The features" refers to anything at all related to objects.

* Begin work on creating the ObjectTable object chain.

This will eventually be the underlying mechanism which holds the
defintion of all objects. The first step is to create the structure, the
second step is to be able to fill it in with native classes. There are
several other steps after this to make it useable, but these are the
steps necessary to make it complete.

After it is complete, the next step will probably be to just allow
native types to provide properties, but this will require quite a bit of
work, especially when it comes to ensuring that all the necessary info
is available through annotations, and parameters are all compatible with
Mixed, etc. However, finishing the ObjectTable will be a great first
step.

* Add ObjectHelpers. Support generic parameters in NonInheritsImplements.

Begin work on a POJOConversion framework, which will allow the exposed
methods in a native object to use POJOs, while transparently in the
background, the values are converted to and from Mixed values for use in
user code. This should make the code much easier to write, and much
faster as well. It may also assist in the future if there is a need to
map third party classes into the ecosystem.

* Expand on the object model in the design docs

Also write the design docs for dependency injection.

* Wrote more on topics, write some of the standard library.

Several more sections have been added to the objects design doc, as well
as a small section about manifests and annotations. In addition to this,
a new cmdline tool, new-type has been added. This allows quick and
correct creation of classes (and types) to be created. Additionally,
many classes in the standard library have been added, though they are
not yet compilable, since the compiler does not support this. However,
they serve as a great base to begin working against for the compiler
changes, and also to see larger segments of code using the designs
described in the design doc. There may be some (unintentional) compiler
errors in the code, but those will be identified later in the process.

Primarily the MethodScript code that is written focuses on supporting
the manifest file, but several pieces of supporting code have also been
written, all primarily annotations. The code is subject to be moved
around and refactored though.

Additionally, a workspace has been created for Visual Studio Code to
more easily edit the methodscript directory.

* Continue work on ObjectTable