Continue working on the functional approach to object creation.
define_object and new_object are new functions, which will serve as the
foundation of the object management system. They will not be used
directly, the compiler will support specific syntax sugar for these
methods, but the methods themselves are separated from the compiler.
Currently, defining the class works, (it successfully puts an empty
class into the ObjectDefinitionTable) though most of the rest of the
system works with the NativeTypeList, rather than the
ObjectDefinitionTabble, which is a problem that must be corrected before
most things will recognize user defined classes. new_object does not
work yet, but progress has been made at least. Additionally, FileOptions
has added a copyright and license field, as well as an UltraStrict
compiler option. This will be used in the future to make a super
pedantic mode, and will be used in all native classes to ensure perfect
compliance with all standards. It is not necessarily recommended for
use in user code, but none-the-less will be available and advertised.