Revert changes in ArgumentValidation.
This proved to have devastating consequences to the performance of
the system, and for the time being, reverting this is the best course
of action. Something downstream of isInstanceOf is taking far too long,
and for a method that would be called extremely frequently, this needs
to be heavily optimized before it can be widely used.