Create UnqualifiedClassName
Elements that are not yet fully defined, but can't be defined yet
can now use UnqualifiedClassName. This can't be used during normal use,
but can be used in the middle of compilation. This solves the bootstrapping
problem that is faced when a class definition extends/implements a class
that hasn't been defined yet, and so thus cannot be fully qualified.

Once the classes have all been initially defined, we can go back through
the list of ObjectDefinitions and call qualifyClasses, which will fully
qualify the previously unqualified class names, and throw an appropriate
exception if any of them are unknown.

Also, rename Objects to ObjectManagement, to match the theme.