Do code maintenance on CommandHelper and AliasCore.
A lot of this code was ancient and had built up cruft over the years. So, this cleans up a number of unused or redundant bits of code (especially in AliasCore.alias(), which is heavily rewritten to be more readable and maintainable), moves logging to the plugin logger where appropriate (avoids issues like double tagging in Spigot), and improves line lengths.