Comment out broken Mixed argument check for now.
This code doesn't appear to have worked for over seven years and is currently just slowing down script evaluation time, particularly when isInstanceOf() was added in this workhorse section of the code. It might not even be needed, or should possibly be done at compile time only. In any case, we haven't depended on it for several years.