Add cmdline support for Windows.
This temporarily removes GUI based support for .ms files,
since that was unreliable. It will be re-added at a later date.
This adds support for a mscript command in cmd.exe, as well as
a PowerShell module, Invoke-MethodScript. The command syntax for
these are slightly different, for the cmdline tools, you must use
the -Tool switch, rather than the -- switch, but otherwise, it is
identical to the unix based system.