Begin work on site translation framework.
This is still not done, and the generated memories are not currently
useable. However, this is a good enough of a stopping place for now.
Still to do is to parse tables, which will solve most of the remaining
problems on the translation memory side. Still to do is write out the
individual pages memories, as well as of course set the site itself up
to start using the translations, but this is a good start on the
process.