Generation of the memory files is nearly complete.
Still remaining to do is to finish the segmentation. After that, the
remaining total tasks are to write a localization interface for
localizers to use, and of course make the site use the translations.
Finally, we need to go through all the summary file translations, and
decide if a translation is eligible for machine translation or not.
Strings that have no values that need to be kept in english can be
marked as such, and all locales will accept the machine translation for
that string. Then, a tool to actually do the machine translation for the
eligible values needs to be written.