L10N Interface is useful now!
Changing settings and saving works now, which is enough to move on to
the next step, which is making a local deployment of the site use the
translations. There are still a lot of steps to go back and finish up,
but I want to get this validated first, so I'm moving on.