The basic features of the l10n ui work, segments are replaced
The basic features of the UI are all completed now, including the
machine translation integration with azure. There are still a few minor
features that need implementing, particularly the segment search
feature, the segment filters, and the help file needs to be brought up
to speed.

The basic replacement logic for the site is implemented as well, and
should start working if pointed to a translation DB in the settings,
however, the replacement is still not working in several edge cases.
Additionally, the site deployment pipeline will need to be updated
with the translation DB link once it exists. The logic to select the
language needs to be implemented as well.

Futher, there are still lots and lots of segments that shouldn't be in
the DB, and so I need to add some more filters to get rid of several
hundred more. By and large however, the segmentation logic is complete,
and anyways bogus segments are a lower priority. The main blocker at
this point is the re-application logic on the site.

There are a few other things on the website that bypass the segmentation
logic currently, particularly menu items and other strings that are
defined in frame.html. I think the best approach here is to just
manually create entries for them. The UI can probably be modified to add
a hidden UI to add new segments, such that it hampers people enough to
not casually add segments, but still allow the rare cases where that
makes sense to easily and consistently occur, and take full advantage of
the UI's validity checking.