Add update operation and fork operation
Also created a framework to easily wrap github api functionality. The
update operation is accessed from the File menu, and essentially is a
git pull on the local repo. The fork operation forks the master repo on
github, clones it locally, then loads it in the UI. Next up is creating
a mechanism to make commits and then make a PR back to the master repo.