Add a check for environment in the compiler
This considerable change now adds a distinction between embedded and
core functions. Embedded functions have a dependency on
CommandHelperEnvironment, and core functions are available everywhere.
Now the compiler makes a distinction between the two, allowing for
compile time errors for functions that aren't available in that
environment. However, there is a need in some cases to compile against
other environments, so the tooling and compiler infrastructure has been
modified to support the notion of a target environment, not the current
runtime environment, hence the size of this change. Additionally, some
tooling, such as the verify tool takes advantage of this distinction, to
allow users to specify what environments they actually want to compile
against.