Give the Main class some TLC
All cmdline tools are in the new @tool format, and the docgen for the
tool should be back to being complete, and should also be more accurate
as well. The verify tool has a few more options added now, particularly
it was too difficult to load extra extensions previously, but with the
new format, this was easy to do, and is now done.