Begin the basic structure of the LanguageServer
I have also created a basic json->java object->json mapper utility. It
has very basic support, but it is in some ways an improvement above
gson, and allows for generic support in a bit of an easier way. Due to
type erasure, there are still serious problems with really supporting
this unfortunately, but this is probably the best that can be done. One
remaining task is to implement enums support, before the work on
implemenation of the language server protocol can begin in earnest.