Disable ElementSignature for now
There is an unaccounted for signature type, if a method returns a non
static inner class, in an outer class which contains generics, this
breaks processing, and causes a StackOverflowError. Therefore, since
this functionality isn't directly needed, I'm just disabling it for now.
Before it can be re-enabled, this bug will have to be fixed.

The binary name that it fails on is something like:
"()Lobj/Name<TT;>.InnerClass"