Continue work on the language server.
The basic RequestMessage framework is complete, and we succesfully read
in most of the initialize method parameters. It's quite a large method
though, and converting the typescript to java is tedious, so there is
still more work to be done before the request itself is finished, but
the framework is complete and works well, so it's just a matter of
finishing the conversion. After that, I need to work on the framework to
handle responses, and that's the next big task.