Enable compiler warnings and fix but in json-api
Compiler warnings were just printed to the log, but couldn't be returned
in a usuable format by the language server. Now, in addition to logging,
the warnings are added to the compiler environment, and can be retrieved
later. Unfortunately, this meant that the lexer needed access to the
Environment, so a new parameter needed to be introduced, and that caused
lots of code changes.

Additionally, older extensions that lacked the MSExtension annotation
would cause issues when loading, because that caused the json-api
command to fail. This fixes it so that if the identifier would be
missing (thus causing a NPE), we generate an identifier based on the jar
name. The VSCode Extension only loaded the language server after the
json-api call completed successfully, so this fixes that issue as well,
if the user has older extensions.