Use proper document models on change
Previously, we just always read from the file system to get the document
truth, but that caused issues, and also required users to save to get
the compilation hints. Now, the document is only read from disk if the
client isn't managing it, otherwise, it reads it from the internal cache
which is managed by the client for open files.