Stop compiling on change, and do compile on save only.
A delay needs to be implemented, otherwise for larger source files, this
causes useability issues, since it would attempt to be compiled once per
character typed. Creating a delay is necessary, so it waits until no
changes have been made for a delay, and only then compiling once. In the
meantime, it will just have to be compile on save.