Add support for method line numbers (ish) in ClassDiscovery
The line number the method declaration is on isn't actually available in
class files, so it's not possible to get that directly, but we can get
a relatively close proxy, the line number of the first executable
statement. This is good enough for most purposes, which is to direct
users attention to roughly where the code is defined. This will
hopefully be good enough for most purposes, because it's the best we can
do without actually parsing java source code.