Support grouping in preference files
Add a few groups to the preferences.ini file also, to take advantage of
this new feature. Groups are sorted alphabetically, and within groups,
items are sorted alphabetically, though within the group, another sort
order may be given to override this. This allows for more precise
sorting of values, and for larger preference files, makes them much
easier for users to read. Values without a group will be placed at the
top of the file.