Use exe wrapper on Windows
There was a bug with the old batch->powershell->java wrapper when it
came to quoted arguments. To fix this became a rabbit hole, and it ended
up being easier to just revist the C# wrapper. This removes the
ClickOnce installation, so this is just a simple wrapper executable.
Still need to revisit the installation mechanisms later to re-add double
click support in windows, but for the time being this is good enough as
a replacement in cmdline.