Add dark mode to the website.
Actually, this is more like "dark theme" because the mechanism is a
generic theming mechanism, which should be pretty straightforward to add
other themes to down the road. Simply add a new theme css file in
resources/css/themes, and add a link to it in the place specified in
frame.html (ensuring that the data- attribute is there). The rest should
figure itself out on its own.