Exclude dependency conflict.
The conflict causes NoSuchMethodError for Spigot methods in BukkitMCItemMeta. Relocation alone can't resolve this. Initially included for lsp4j, com.google.common.collect doesn't appear to be used in the functionality LangServ is using at the moment, and so it should be safe to exclude it for now. This also reduces the jar size.