Add BlockUntilGC method.
This wraps System.gc(), but provides a way to actually ensure that a
garbage collection has actually run. This is no still guarantee that a
given object has been garbage collected, but is as good as it can get
from a java code perspective. There are a few places where System.gc()
has been changed to use this new method instead.