Add qr_code
This doesn't currently build unless you manually clone and build the
qrcode-terminal project. In theory, this should already be published in
the github package registry, and will build if you set up a connection
to github, however, building should not require a github account, as
this is intended to be a public project. Given that github pacakages is
still in beta, I will wait to merge this into master until I can get the
qrcode-terminal project working correctly.