Add static analysis framework
Functions that have scoping behavior or functions that make declarations / references to variables/procs/etc should override the linkScope() method.
Variable references are already checked at this point, but declarations are not.
Comes with a boolean to do or don't do static analysis. This is necessary to run optimization tests that are semantically incorrect.