First version of two-pass scope graph analysis
Change static analysis to first generate the scope graph, and later do analysis on that scope graph. This allows static includes (include()) to be properly supported, even when containing cycles ((in)directly including themselves), and it offers easy support for checking procedure names/signatures in the future.
Note that this commit runs on my code base, but still needs testing, finetuning and a code cleanup.