Add support for telemetry
This branch adds the ability to send telemetry data to Azure Application Insights. This proxies through apps.methodscript.com however, so in addition to the basic telemetry framework, this also includes the swagger framework, which both the server and the client have autogenerated code for. See the LadyCailin/apps.methodscript.com repo for the server side of this. Telemetry is diabled by default, but nags the user to enable it where relevant.