Add rewrite step between parsing and static analysis
Add AST rewrite step that should be used to rewrite the AST that is received from the parser to a valid executable AST. This is necessary because the parser leaves some weird terms in the AST that need to be rewritten. These rewrites happened in optimization methods, while they were not actually optimizations. Since static analysis has to happen on a valid AST before optimization, this change has been made.
Note that only the array_get and array_push rewrites are moved in this commit.