Allow procs to use not-yet-defined procs
Allow procedures to use any not-yet-defined procedures, as long as those procedures are defined when the procedure is called. This allows for much freedom using procs, while still ensuring that used procedures are defined at runtime. There is one catch, being that when a procedure is not used, any invalid procedure calls within it will not be checked.