Instantiate GlobalEnv with cmdline and interpreter setting
Since we always know whether the environment is in cmdline and interpreter mode, it makes more sense to pass these variables to the GlobalEnv constructor such that we can never forget to set them. This also means that we don't have to come up with some default value for them.