[Breaking] Change cmdline and interpreter base dir
Change the cmdline base directory to be the currently running script file container. Change the interpreter base directory to be the GlobalEnv root directory (setable through cd()). Reasoning behind this is that using cd() in a shell is a nice feature, but in a script it is making much less sense. This change also makes cmdline mode scripts and MC server scripts behave the same in terms of usage of relative paths (in include(), read(), etc), and therefore more compatible with each other. Static analysis is also much more doable if the base directory is actually a fixed value.