Create standalone Windows installer for cmdline.
This uses NullSoft Scriptable Install System to create the executable,
which is fully contained and downloads the necessary files, including
Java, from the internet. This prevents the installer itself from needing
to be updated on each update. Eventually, the installer should be
signed, but for now it's unsigned, causing various scary warnings upon
installation. In any case, the installer code itself works, as well as
the installer.