Fix exception caused by unknown compass meta world
Fix NullPointerException that occurs when getting the location CArray from MCLocation in ObjectGenerator, in the case that the world is `null`. This case triggers when loading BukkitMCCompassMeta through the CreativeInventoryEvent when a player loads a compass with target meta data that refers to a world that does not exist on the current MC server (causing `CompassMeta.getLodestone()` to contain a `null` world).